requirements.txt
asgiref==3.2.10
autopep8==1.6.0
Django==3.1
django-ckeditor==6.2.0
django-js-asset==2.0.0
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
